requirements.txt
boto==2.48.0
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.5
django-ckeditor==5.3.1
django-cors-headers==2.1.0
django-filter==1.1.0
django-js-asset==0.1.1
django-nested-admin==3.0.21
django-nested-inline==0.3.7
django-rest-framework==0.1.0
django-storages==1.6.5
djangorestframework==3.6.4
gunicorn==19.7.1
idna==2.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
python-monkey-business==1.0.0
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
